将升级器同步到全局，现在可以继承 `BackwardCompatible{?}MetadataJsonConverter` 来获取统一的升级器功能

+++ RhythmBase.BeatBlock/BeatBlock/Components/Metadata.cs
@@ -142,7 +142,7 @@ public class Chart :
     /// <summary>
     /// Gets or sets the difficulty of the chart.
     /// </summary>
-    public float Difficulty { get; set; }
+    public int Difficulty { get; set; }
     /// <summary>
     /// Gets or sets the display name of the chart.
     /// </summary>
